requirements.txt
certifi==2020.6.20
websocket-client==1.2.1
